Add files via upload
Unfortunately my GPU crashed before I could finish training and on top of that my computer ran out of battery and shut down before I noticed, meaning that the output of the training was also lost. Hopefully my code offers enough explanation to assess whether or not it would have produced faces in the end.